requirements.txt
PyYAML==6.0.3
readchar==4.2.1
